* 'lab7-kswapd' of github.com:etteerr/KernelProgrammingLab:
  Uncommented unswappable error
  Fixed swapout + Alot
  Changed kswapd order and added used condition to get_mem_rss
  More agressive yielding, print elaboration
  Changed atomic order for swapin queue
  Made bytes MiB
  Sped up mempress
  General fixes
  Print statement changes and atomic order in swappy
  handle_pf_pte now checks if there are 2 pages available
  Added items remaining print for swappy
  Removed "cannot allocate, yielding" message
  Removed newhead message Added appropriate yields in kswapd_service
  Optimized kswap_service loop
  General swappy and reverse_pte fixes
  Changed iteration of kswapd_service Fixed iterators not resetted in try_swap Changed swap_call in try_swap
  Made dprinfs encapsulated
  Kernel panic is now BSOD

# Conflicts:
#	kern/kswapd.c

+++ kern/sched.c
@@ -118,7 +118,7 @@ void sched_yield(void)
     /* Iterates over envs, starting at curenv's index, wrapping
      * around NENVS to 0, and from there up to curenv's index. */
     for(i = 0; i < NENV; i++) {
-        env_i = shared_sched_get_next_iter() % NENV;
+        env_i = shared_sched_get_next_iter() % (unsigned long long int)(NENV);
         idle = &envs[env_i];
 
         /* Try to acquire lock (tests if runnable) */
